# OpenAPI.yaml
openapi: 3.0.0
info:
  title: MY SWEET SKRIPSI
  contact: {}
  version: "1.0.0"
servers:
  - url: http://localhost:8000/v1/api
  - url: https://be-permission.up.railway.app/v1/api
    variables: {}
paths:
...
  /santri/byid/{id}:
    get:
      tags:
        - For Data Student
      summary: GET DATA SANTRI BY ID
      description: ""
      operationId: GETDATASANTRIBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful Get data santri
          headers: {}
          content: {}
        "400":
          description: Error or not Success get data santri
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
  /kriteria/byid/{id}:
    get:
      tags:
        - For Data Kriteria and Sub-kriteria
      summary: GET DATA KRITERIA BY ID
      description: ""
      operationId: GETDATAKRITERIABYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful Get data kriteria by id
          headers: {}
          content: {}
        "400":
          description: Error or not success get data kriteria by id
          headers: {}
          content: {}
        "500":
          description: internal server error
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security: []
...